[#19178] YSQL: Invalidate response cache in case of discard temp tables
Summary:
In case of removing table via `DROP TABLE` command YSQL increments catalog version because this command alters/removes some existing data. But when temp tables are removed via `DISCARD ALL` command (or when connection is closed) catalog version is not incremented, because catalog data is modified only for tables which is used by current connection only. This approach helps to avoid catalog cache refreshes on all active connections in case connection with at least one temp table is closing.

But this approach produces abnormalities in case response cache is enabled.
The following scenario shows this abnormalities:
```
- start new connection conn1 with ysqlsh
  CREATE TEMP TABLE tmp(k INT);
  CREATE TABLE t(k INT);DROP TABLE t; -- for catalog version bumping

- start new connection conn2 with ysqlsh

- close conn1

- wait for 3-5 secs

- start new connection conn3 with ysqlsh
  CREATE TEMP TABLE tmp(k INT); -- fails with `ERROR:  relation "tmp" already exists`
```

Possible solution is to increment catalog version in case of temp table is removed (because catalog version is a part of the catalog response cache). But this will leads to catalog cache refresh in all active connection in case of closing connection with temp tables.

Better solution is to explicitly invalidates response cache without incrementing catalog version. For this purpose in the new DDL mode `YB_DDL_MODE_SILENT_ALTERING` is introduces. This mode indicates that current DDL modifies  some existing data in catalog, but without catalog version increment. On committing this kind of DDL `PgClientSession` will invalidate catalog response cache for particular DB.
Actual algorithm of the cache invalidation is the following:
- before committing DDL transaction with the `YB_DDL_MODE_SILENT_ALTERING` mode the response cache for particular DB is marked as disabled. (**Note:** cache is disabled for a group of DBs (which are placed inside single bucket) and not for an individual. see the description of the `pg_response_cache_num_key_group_bucket` flag for details)
- while response cache on particular DB is disabled any request related to this DB will no use data from the cache. All request will go to a master instead.
- after DDL commit is completed (with any status) response cache for particular DB is marked as enabled

**Note:**

In context of this diff new metric is introduced

`pg_response_cache_disable_calls` - contains total number of response cache disabling (for all DBs)

In context of this diff new flag is introduced

`pg_response_cache_num_key_group_bucket` (default value is `512`) - specifies the number of buckets to distribute DB IDs (DB ID is used as a key group within the response cache).
Jira: DB-7975

Test Plan:
New unit tests were introduced

```
./yb_build.sh --gtest_filter PgCatalogPerfTest.ResponseCacheInvalidationOnConnectionWithTempTableClosure
./yb_build.sh --gtest_filter PgCatalogPerfTest.SameTempTableCreationWithResponseCache
./yb_build.sh --gtest_filter PgCatalogPerfTest.ResponseCacheInvalidationOnDiscardTempTables
```

Reviewers: fizaa, myang, kfranz

Reviewed By: fizaa

Subscribers: yql, ybase, bogdan

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D28724

+++ src/yb/yql/pggate/pg_session.h
@@ -253,7 +253,7 @@ class PgSession : public RefCountedThreadSafe<PgSession> {
       ForceNonBufferable force_non_bufferable = ForceNonBufferable::kFalse);
 
   struct CacheOptions {
-    uint64_t key_group;
+    uint32_t key_group;
     std::string key_value;
     std::optional<uint32_t> lifetime_threshold_ms;
   };
